[PK-11]
Add tab "wspolnoty parafialne" in Parafia Tab
